Convert rootfs symlinks to be relative
This isn't necessary once we're inside of the sandbox, but if we try to
do something like mount `/var/run/docker.sock` within the sandbox, and
it turns out that `/var/run` is a symlink to `/run`, the sandbox will
try to dereference the symlink and fail.  So we try to make things
consistent everywhere.