Add macOS seatbelt-based runner
Seatbelt is the macOS sandboxing mechanism in use by the internal macOS
app sandbox, as well as projects such as Chrome to provide OS-level
access controls on subprocesses.  We are using it here to provide a
read-only view of the world, with large-scale access only to certain,
particular folders.

We have a plan to allow for secrets management a la `cryptic`, whereby
we copy the secrets in to a particular path that is visible to the
sandboxed process, but then delete it in the environment hook if it's
not a privileged run.  This is yet to be completed.