Add `num_cpus` config option
This builds cgroups with pinned cores, for better worker isolation (and
to teach `rr` to not always pin to the same first few CPUs).  This is
implemented by creating wrapper scripts that generate cgroups and pin
agents to specific cores.  This has the added benefit of completely
disallowing a poorly-behaved worker process from taking over other
worker's cores.