Add a `stop_timeout` of 120 minutes and use `kill_mode` `mixed`.
This will allow us to gracefully stop our buildkite-agent processes,
instead of brutally murdering them and interrupting jobs every time we
have a new configuration change.  This will send `SIGTERM` to the
sandbox process, and then wait up to two hours before sending a
`SIGKILL` to the entire process group.

For this to actually work, we need to wait until sandbox responds to
signals properly, which is pending this PR [0] getting merged.

[0] https://github.com/staticfloat/Sandbox.jl/pull/90