Add `${HOME}/Library/Preferences` directory
Without this directory, `security list-keychain -d user -s ${keychain}`
will silently fail to add the keychain to the path.  This took a lot of
debugging, including turning off SIP to be able to run `dtruss`, so I'm
a little grumpy about it, but on the other hand, it's great to have
DTrace working again on Apple Silicon.