If the pristine OS image gets modified, we need to update our cache image too
This occasionally didn't work properly because when we rebuild the image
while the agent is running, it is writing to its cache image constantly.
Then, when the build here is done, we reboot the agent VM, and its cache
image gets a newer timestamp than the newly created one.  This instead
locks everything to the timestamp of the pristine image, which does not
get updated.