Update buildkite agent release
The buildkite servers seem to have started throwing 502 errors on
disconnect now and then.  They merged a workaround in v3.39.0 [0], so
let's update and see if that improves our agent reliability.

This also updates the `update_agent_rootfs_bindings.jl` script to
automatically bump the agent release on macOS as well.

[0] https://github.com/buildkite/agent/releases/tag/v3.39.0
    X-ref: https://github.com/buildkite/agent/pull/1761