Alter KVM XML machine template
It appears that FreeBSD 12 does not support the Q35 machine very well,
and we should instead use the i440FX machine [0].  This has the
consequence of not supporting PCIe and whatnot, so I had to make some
big changes to the XML template.  Luckily, I basically just had to
delete a bunch of stuff that I think Windows needed, so this is nice in
that it decreases the amount of configuration that nobody understands
anyway.

This was all necessary because FreeBSD was failing to find the virtio
drive after booting under KVM for the first time.  This makes sense, as
Packer defaults to a `pc` machine type, which is an alias for `i440FX`.

[0] https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=236922