Wrap the rc.d buildkite-agent invocation in `su` (#48)
According to a discussion on the FreeBSD forum
(https://forums.freebsd.org/threads/how-to-run-rc-d-service-as-specific-user.86879/),
the specified user may not actually be the one who owns the
`buildkite-agent` process. While buildkite seems to be configured to run
jobs as said user, there seems to be some kind of configuration mismatch
since the Julia tests claim they're being run as root. We can try to fix
this by wrapping the command to launch `buildkite-agent` in `su`.