[freebsd-kvm] Ignore nonzero exits from `buildkite-agent`
It seems that when the `buildkite-agent` process exits with a nonzero
code, the `buildkite_start` function in the rc.d script returns before
it gets to `shutdown`. This leaves VMs with failed jobs just sitting
around, since `buildkite-agent` is no longer running and nothing else is
there to shut them down.