Load FreeBSD virtio drivers at boot time
We've specified virtio as the network adapter for KVM and it works but
we get occasional timeouts from httpbingo and GitHub while running the
Julia test suite, which suggests a possible network setup issue. FreeBSD
has virtio kernel modules that can be loaded on boot so we can try
loading those to see if they help. See `man 4 virtio` for where these
flags came from.