Lock to buildkite agent v3.39.0 on FreeBSD, and disable git-mirrors
Unfortunately, our FreeBSD agents seem to have a hard time cloning with
`--mirror`, something about the request borks github and we end up
timing out, or getting a bad response, or something.  By disabling the
`--mirror` optimization, we are able to get things working.  I suppose
we'll try turning it on again in the future.  This used to work in
FreeBSD 12, so I suspect something like libcurl v8.1 is causing an issue
here.