Use persistence, but clean it out
If `persist=false` is set, sandbox will create a `tmpfs` and mount the overlay workdir wtihin that tmpfs.  This works, but then fails horribly when `apt` tries to rename things from `/tmp` (which is mounted on our host system) to the rootfs (which is backed by the overlay).
The current workaround is to force the persistence directory to be backed by our actual filesystem as well, and then delete the persistence directory as part of our job cleanup.