Reconfigure DNS resolution on FreeBSD VMs
We've been experiencing a failure in Julia's Downloads tests on FreeBSD,
reproducible only in our VMs: the request to https://domain.invalid hits
the 30-second timeout set by default in Downloads because it takes a
full 30 seconds for the resolver to report `NXDOMAIN`. I have no idea
why, and I can't even reproduce it in a FreeBSD VM on my local machine.
Absent a real fix, we can apply a band-aid in the meantime: prefer
Google's nameserver in the DNS resolver configuration. In the debugging
VM on amdci6, this allows the Downloads tests to pass succesfully.