Windows: Reorganize virsh start-up sequence. (#109)
I think we were running into the issue that the first ExecStop command,
shutting down the VM, failed because of the active shutdown initiated
by nssm, resulting in the virsh destroy not executing. This in turn
caused the qcow2 images remaining owned by the libvirt user, resulting
in subsequent jobs failing to copy and overwrite the disk images.
Avoid this by only unconditionally destroying the image. This should
hopefully not result in corrupt images, as it's assumed the cache disk
is not being written into anymore as part of the shutdown.