Fix persistence directory selection
Sandbox requires us to carefully find a persistence directory that
supports all the features that overlayfs requires.  It has a helper
function, called `Sandbox.find_persist_dir_root()` that can take in some
hints, so let's use that function and plumb through the ability for the
user to select a proper location to hold all the upper and work
directories.  In the case of Yggdrasil, we place this intentionally so
that our rather large upperdirs are located on a drive with sufficient
space.